refactoring report templates,merging child class buildings and roads
