+++ safe/impact_template/land_cover_report_template.py
@@ -68,8 +68,6 @@ class LandCoverReportTemplate(TemplateBase):
         message.add(self.format_impact_summary())
         message.add(self.format_action_check_list())
         message.add(self.format_notes())
-        if self.postprocessing:
-            message.add(self.format_postprocessing())
         return message
 
     def format_impact_summary(self):
